ref: Separate main.go into a cli builder (#49)
This commit looks fairly busy, but there is very little new logic.

I have broken up the huge `main.go` into individual command files under
`pkg/command`. New code can be found in `pkg/flags`, `pkg/defaults` and
`pkg/config`. 95% of the content of `pkg/command` is lifted straight
from what was originally in `main.go`, the only difference is how the
app is constructed.

Each command now loads flag values from the context into a Config object
which can be used by the command action. Each command action is a
function which will be individually unit tested. Previously we were
testing all the permutations in integration tests, which was just too
much of a pain to do. Now we should be able to use a mocked client.

There are a couple of other things that I would like to do:
- Add tests around the action funcs
- Have another layer between the actions and the client
- Have a dialFunc which is passed into the action func. This will be
  helpful when we add TLS and auth so we don't have to repeat the set up
  of creds for each command

But I am leaving that for now.