ref: Add Client builder for generic abstraction
This adds a `ClientBuilderFunc` to the command Config option, and moves
the create logic back into the `CreateFn` and out of the `client`.

This gives a nicer abstraction layer for testing, as well as a smaller
shim over the real flintlock client.

There is more scope for refactoring the busier commands (eg `delete`),
but this make things easier for now.